requirements.txt
Django>=5.0,<6.0
django-allauth>=0.63
Pillow>=10.0
django-taggit>=6.0
stripe>=5.4
